Extract some substitutors, move compilers to separate folder
