+++ Source/Dafny/AlphaConvertingSubstituter.cs
@@ -9,10 +9,10 @@ namespace Microsoft.Dafny
   /// case where "receiverReplacement" is given as ImplicitThisExpr (but no attempt is made to substitute for all ImplicitThisExpr's in
   /// "receiverReplacement" and the range of "substMap").
   /// </summary>
-  public class AlphaConverting_Substituter : Substituter
+  public class AlphaConvertingSubstituter : Substituter
   {
     ISet<string> namesToAvoid = new HashSet<string>();
-    public AlphaConverting_Substituter(Expression receiverReplacement, Dictionary<IVariable, Expression> substMap, Dictionary<TypeParameter, Type> typeMap)
+    public AlphaConvertingSubstituter(Expression receiverReplacement, Dictionary<IVariable, Expression> substMap, Dictionary<TypeParameter, Type> typeMap)
       : base(receiverReplacement is ImplicitThisExpr ? new ThisExpr(receiverReplacement.tok) { Type = receiverReplacement.Type } : receiverReplacement, substMap, typeMap) {
       Contract.Requires(substMap != null);
       Contract.Requires(typeMap != null);
